requirements.txt
Django
google-generativeai
pillow
requests
python-dotenv
gunicorn
whitenoise